Added postprocessing to the Relax Subworkflow. Now the relaxed structure is added to a specified location of the spec. Magnetic moments are read from the OUTCAR and added to the structures sites.
FT_PrintSpec should be probably removed after testing.
It should be checked that the spec is then transfered to the
next Firework correctly. Currently relaxing
with SCAN only works if a k_distance is set (means that the
generalized kmesh is used). This can maybe be fixed by looking
at custodians 'metagga_opt_run' and why it does not work without
a KPOINTS file even if KSPACING is set in the INCAR.

Docstrings need to be written.