Bug fixed regarding updating of the spec between Fireworks.
Apparently not only the last Firetask is important, but calling
update_spec = X in an erlier FWAction influences the spec in the
next FW. Switched to mod_spec '_set' for the FT_CopyInSpec Firetask
and suddenly everything works.

TODO: Check if this mod_spec method can be used in other Firetasks
as well.