MPInterfaces get_aligned_lattices in version 2020.6.19 (at least) runs into crashes for max_angle_diff > ~1.5, at least for certain interfaces. A match is found, but than there is a LinAlgError("Singular matrix") error in forming the matched slabs? I added two lines to increase max_angle_diff to 1.5 if it is smaller. This is not a great solution obviously. I also changed the default in FT_CheckInterfaceParamDict to 1.5. It was 1.0 before.
Another thing was to change the Slab generation in FT_GetRelaxedSlab so that 'scale_factor' is now a list
of lists and not a np.array. The numpy array was leading to an error when putting the Slab in the DB because
it could not be json-serialized. This should have no other effects, but we should watch out!