Modified GetCustomVaspRelaxSettings to include an 'interface_z_relax' allowed type which sets selective_dynamics ([False, False, True] for each site) for the structure. This directly carries over to the vasp input set.
Added the two helper functions StackAlignedSlabs and ReCenterAlignedSlabs. The latter is used
to center the aligned slabs that come out of the MPInterface function get_aligned_lattices
so that they are centered around z=0. A strange thing here is that the aligned slabs do not
have the exact same lattice! That should be further investigated.
StackAlignedSlabs then allows to form an interface using the centered slabs. An optional
shift vector in cartesian coordinates can be used to sample different high symmetry sites
or change the distance between the slab for the PES and PPES.