Apparently pymatgens MPScanRelaxSet defaults now to r2SCAN, a very recently defeloped improvement of the 'clasic' SCAN functional. However, this is not yet implemented in VASP, as is also acknowledged in the documentation:
"This functional is currently not officially supported in VASP.
Source code may be obtained by contacting the authors of the manuscript in the References section.
The original SCAN functional, available from VASP 5.4.3 onwards,
maybe used instead by passing user_incar_settings={“METAGGA”: “SCAN”} when instantiating this InputSet.
r2SCAN and SCAN are expected to yield very similar results."

Since we should not assume that the users of our tools will
download and compile modified VASP code, we will stick for now
with the original SCAN version.