This should actually have been in branch 26-PES_SW-rewrite, but since it was a rather small and easy fix I decided to not branch off again from development and reopen issue 26.
This fixes a problem I had with copying the results to a remote machine
which did not work properly with the built in FileTransferTask of
Fireworks. For remote copy, now we use a ScriptTask which simply runs
a bash scp command. The docstrings of the necessary functions have
been extended and include a warning to copy files only cautiosly and
if the prerequisites are met.