Modified dynamic_relax_swf in triboflow.workflows.base
to accept a third input parameter "add_static". If set to
true, a static calculation is added after the relaxation
runs. This is done to get accurate energies, and uses ISMEAR=-5.
To enable WAVECAR copy, StaticFW of atomate had to be modified slightly.
Since we modify this anyhow to enable vdw_kernel copy, this is no big deal.

A test was written to test this functionality as well.